bugfix in 3D plotting for less than full dimensional sets, getting rid of recursive best-dive implementation

+++ src/SetOperations.cpp
@@ -853,7 +853,7 @@ namespace ZonoOpt
         if (Z.zero_one_form) Z.convert_form();
 
         // dimension variables
-        const int n_cons = H.rows();
+        const int n_cons = static_cast<int>(H.rows());
         const int n_slack = direction == '=' ? 0 : n_cons;
 
         // re-used matrices
